refactor: controller, model, repository 패키지 생성
