+++ src/main/java/gift/controller/ViewController.java
@@ -1,5 +1,6 @@
-package gift;
+package gift.controller;
 
+import gift.model.Product;
 import java.util.List;
 
 import org.springframework.stereotype.Controller;
