+++ src/main/java/gift/model/Product.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.model;
 
 public record Product(long id, String name, long price, String imageUrl) {
