+++ src/main/java/gift/repository/DatabaseInitializer.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.repository;
 
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
