+++ src/main/java/gift/repository/ProductRepository.java
@@ -1,5 +1,6 @@
-package gift;
+package gift.repository;
 
+import gift.model.Product;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
